MDEV-20004 Move Field_geom from field.cc to sql_type_geom.cc
