+++ sql/field.cc
@@ -8781,242 +8781,6 @@ longlong Field_blob_compressed::val_int(void)
 }
 
 
-#ifdef HAVE_SPATIAL
-/* Values 1-40 reserved for 1-byte options,
-   41-80 for 2-byte options,
-   81-120 for 4-byte options,
-   121-160 for 8-byte options,
-   other - varied length in next 1-3 bytes.
-*/
-enum extra2_gis_field_options {
-  FIELDGEOM_END=0,
-  FIELDGEOM_STORAGE_MODEL=1,
-  FIELDGEOM_PRECISION=2,
-  FIELDGEOM_SCALE=3,
-  FIELDGEOM_SRID=81,
-};
-
-
-uint gis_field_options_image(uchar *buff, List<Create_field> &create_fields)
-{
-  uint image_size= 0;
-  List_iterator<Create_field> it(create_fields);
-  Create_field *field;
-  while ((field= it++))
-  {
-    if (field->real_field_type() != MYSQL_TYPE_GEOMETRY)
-      continue;
-   if (buff)
-   {
-     uchar *cbuf= buff + image_size;
-
-     cbuf[0]= FIELDGEOM_STORAGE_MODEL;
-     cbuf[1]= (uchar) Field_geom::GEOM_STORAGE_WKB;
-
-     cbuf[2]= FIELDGEOM_PRECISION;
-     cbuf[3]= (uchar) field->length;
-
-     cbuf[4]= FIELDGEOM_SCALE;
-     cbuf[5]= (uchar) field->decimals;
-
-     cbuf[6]= FIELDGEOM_SRID;
-     int4store(cbuf + 7, ((uint32) field->srid));
-
-     cbuf[11]= FIELDGEOM_END;
-   }
-   image_size+= 12;
-  }
-
-  return image_size;
-}
-
-
-uint gis_field_options_read(const uchar *buf, size_t buf_len,
-      Field_geom::storage_type *st_type,uint *precision, uint *scale, uint *srid)
-{
-  const uchar *buf_end= buf + buf_len;
-  const uchar *cbuf= buf;
-  int option_id;
-
-  *precision= *scale= *srid= 0;
-  *st_type= Field_geom::GEOM_STORAGE_WKB;
-
-  if (!buf)  /* can only happen with the old FRM file */
-    goto end_of_record;
-
-  while (cbuf < buf_end)
-  {
-    switch ((option_id= *(cbuf++)))
-    {
-    case FIELDGEOM_STORAGE_MODEL:
-      *st_type= (Field_geom::storage_type) cbuf[0];
-      break;
-    case FIELDGEOM_PRECISION:
-      *precision= cbuf[0];
-      break;
-    case FIELDGEOM_SCALE:
-      *scale= cbuf[0];
-      break;
-    case FIELDGEOM_SRID:
-      *srid= uint4korr(cbuf);
-      break;
-    case FIELDGEOM_END:
-      goto end_of_record;
-    }
-    if (option_id > 0 && option_id <= 40)
-      cbuf+= 1;
-    else if (option_id > 40 && option_id <= 80)
-      cbuf+= 2;
-    else if (option_id > 80 && option_id <= 120)
-      cbuf+= 4;
-    else if (option_id > 120 && option_id <= 160)
-      cbuf+= 8;
-    else /* > 160 and <=255 */
-      cbuf+= cbuf[0] ? 1 + cbuf[0] : 3 + uint2korr(cbuf+1);
-  }
-
-end_of_record:
-  return (uint)(cbuf - buf);
-}
-
-
-
-void Field_geom::sql_type(String &res) const
-{
-  CHARSET_INFO *cs= &my_charset_latin1;
-  const Name tmp= m_type_handler->name();
-  res.set(tmp.ptr(), tmp.length(), cs);
-}
-
-
-int Field_geom::store(double nr)
-{
-  my_message(ER_CANT_CREATE_GEOMETRY_OBJECT,
-             ER_THD(get_thd(), ER_CANT_CREATE_GEOMETRY_OBJECT), MYF(0));
-  return -1;
-}
-
-
-int Field_geom::store(longlong nr, bool unsigned_val)
-{
-  my_message(ER_CANT_CREATE_GEOMETRY_OBJECT,
-             ER_THD(get_thd(), ER_CANT_CREATE_GEOMETRY_OBJECT), MYF(0));
-  return -1;
-}
-
-
-int Field_geom::store_decimal(const my_decimal *)
-{
-  my_message(ER_CANT_CREATE_GEOMETRY_OBJECT,
-             ER_THD(get_thd(), ER_CANT_CREATE_GEOMETRY_OBJECT), MYF(0));
-  return -1;
-}
-
-
-int Field_geom::store(const char *from, size_t length, CHARSET_INFO *cs)
-{
-  if (!length)
-    bzero(ptr, Field_blob::pack_length());
-  else
-  {
-    if (from == Geometry::bad_geometry_data.ptr())
-      goto err;
-    // Check given WKB
-    uint32 wkb_type;
-    if (length < SRID_SIZE + WKB_HEADER_SIZE + 4)
-      goto err;
-    wkb_type= uint4korr(from + SRID_SIZE + 1);
-    if (wkb_type < (uint32) Geometry::wkb_point ||
-	wkb_type > (uint32) Geometry::wkb_last)
-      goto err;
-
-    if (m_type_handler->geometry_type() != Type_handler_geometry::GEOM_GEOMETRY &&
-        m_type_handler->geometry_type() != Type_handler_geometry::GEOM_GEOMETRYCOLLECTION &&
-        (uint32) m_type_handler->geometry_type() != wkb_type)
-    {
-      const char *db= table->s->db.str;
-      const char *tab_name= table->s->table_name.str;
-
-      if (!db)
-        db= "";
-      if (!tab_name)
-        tab_name= "";
-
-      my_error(ER_TRUNCATED_WRONG_VALUE_FOR_FIELD, MYF(0),
-               Geometry::ci_collection[m_type_handler->geometry_type()]->m_name.str,
-               Geometry::ci_collection[wkb_type]->m_name.str,
-               db, tab_name, field_name.str,
-               (ulong) table->in_use->get_stmt_da()->
-               current_row_for_warning());
-      goto err_exit;
-    }
-
-    Field_blob::store_length(length);
-    if ((table->copy_blobs || length <= MAX_FIELD_WIDTH) &&
-        from != value.ptr())
-    {						// Must make a copy
-      value.copy(from, length, cs);
-      from= value.ptr();
-    }
-    bmove(ptr + packlength, &from, sizeof(char*));
-  }
-  return 0;
-
-err:
-  my_message(ER_CANT_CREATE_GEOMETRY_OBJECT,
-             ER_THD(get_thd(), ER_CANT_CREATE_GEOMETRY_OBJECT), MYF(0));
-err_exit:
-  bzero(ptr, Field_blob::pack_length());  
-  return -1;
-}
-
-
-bool Field_geom::is_equal(const Column_definition &new_field) const
-{
-  /*
-    - Allow ALTER..INPLACE to supertype (GEOMETRY),
-      e.g. POINT to GEOMETRY or POLYGON to GEOMETRY.
-    - Allow ALTER..INPLACE to the same geometry type: POINT -> POINT
-  */
-  if (new_field.type_handler() == m_type_handler)
-    return true;
-  const Type_handler_geometry *gth=
-    dynamic_cast<const Type_handler_geometry*>(new_field.type_handler());
-  return gth && gth->is_binary_compatible_geom_super_type_for(m_type_handler);
-}
-
-
-bool Field_geom::can_optimize_range(const Item_bool_func *cond,
-                                    const Item *item,
-                                    bool is_eq_func) const
-{
-  return item->cmp_type() == STRING_RESULT;
-}
-
-
-bool Field_geom::load_data_set_no_data(THD *thd, bool fixed_format)
-{
-  return Field_geom::load_data_set_null(thd);
-}
-
-
-bool Field_geom::load_data_set_null(THD *thd)
-{
-  Field_blob::reset();
-  if (!maybe_null())
-  {
-    my_error(ER_WARN_NULL_TO_NOTNULL, MYF(0), field_name.str,
-             thd->get_stmt_da()->current_row_for_warning());
-    return true;
-  }
-  set_null();
-  set_has_explicit_value(); // Do not auto-update this field
-  return false;
-}
-
-
-#endif /*HAVE_SPATIAL*/
-
 /****************************************************************************
 ** enum type.
 ** This is a string which only can have a selection of different values.
