+++ sql/item_geofunc.h
@@ -26,6 +26,8 @@
 #pragma interface			/* gcc class implementation */
 #endif
 
+#include "sql_type_geom.h"
+#include "item.h"
 #include "gcalc_slicescan.h"
 #include "gcalc_tools.h"
